Move workout-detail and workout-history
Move workout-detail and workout-history pages under the profile page
