+++ src/app/app.component.html
@@ -10,7 +10,7 @@
       </ion-header>
       <ion-content>
         <ion-list>
-          <ion-item (click)="menuNavigate('/home/workout/history')" button>
+          <ion-item (click)="menuNavigate('/home/profile/history')" button>
             <ion-icon name="bar-chart-outline" slot="start"></ion-icon>
             <ion-label>Workout History</ion-label>
           </ion-item>
