+++ src/app/profile/profile.page.html
@@ -9,7 +9,7 @@
 
 <ion-content>
   <ion-list>
-    <ion-item routerLink="/home/workout/history">
+    <ion-item routerLink="/home/profile/history">
       <ion-icon name="bar-chart-outline" slot="start"></ion-icon>
       <ion-label>Workout History</ion-label>
     </ion-item>
