+++ src/app/profile/workout-detail/workout-detail.module.ts
