+++ src/app/profile/workout-detail/workout-detail.page.html
@@ -1,7 +1,7 @@
 <ion-header>
   <ion-toolbar>
     <ion-buttons slot="start">
-      <ion-back-button defaultHref="/home/workout/history"></ion-back-button>
+      <ion-back-button defaultHref="/home/profile/history"></ion-back-button>
     </ion-buttons>
     <ion-title>Workout Details</ion-title>
     <ion-buttons slot="end">
