+++ src/app/profile/workout-detail/workout-detail.page.scss
