+++ src/app/profile/workout-history/workout-history.page.scss
