+++ src/app/workout/workout-routing.module.ts
@@ -7,21 +7,7 @@ const routes: Routes = [
   {
     path: '',
     component: WorkoutPage,
-  },
-  {
-    path: 'history',
-    loadChildren: () =>
-      import('./workout-history/workout-history.module').then(
-        (m) => m.WorkoutHistoryPageModule
-      ),
-  },
-  {
-    path: 'detail/:workoutId',
-    loadChildren: () =>
-      import('./workout-detail/workout-detail.module').then(
-        (m) => m.WorkoutDetailPageModule
-      ),
-  },
+  }
 ];
 
 @NgModule({
